little bit of cleanup
