+++ ehackathon-server/app.js
@@ -1,3 +1,4 @@
+//modules
 var express = require('express');
 var path = require('path');
 var favicon = require('serve-favicon');
@@ -7,12 +8,10 @@ var bodyParser = require('body-parser');
 var passport = require('passport');
 var RedditStrategy = require('passport-reddit').Strategy;
 var session = require('express-session');
-var crypto = require('crypto');
+
 var config = require('./config.json');
 
-var routes = require('./routes/index');
-var login = require('./routes/login');
-
+//app
 var app = express();
 
 // view engine setup
@@ -35,71 +34,40 @@ passport.use(new RedditStrategy({
   },
   function(accessToken, refreshToken, profile, done) {
     // asynchronous verification, for effect...
-    process.nextTick(function () {
+    process.nextTick(function() {
       return done(null, profile);
     });
   }
 ));
 
-// uncomment after placing your favicon in /public
-//app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
+//express init
 app.use(logger('dev'));
 app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: false }));
-app.use(cookieParser());
-app.use(require('node-sass-middleware')({
-  src: path.join(__dirname, 'public'),
-  dest: path.join(__dirname, 'public'),
-  indentedSyntax: true,
-  sourceMap: true
+app.use(bodyParser.urlencoded({
+  extended: false
 }));
-app.use(session({ secret: 'keyboard cat' }));
+app.use(cookieParser());
+//session/auth stuff
+app.use(session({
+  secret: 'keyboard cat'
+}));
 app.use(passport.initialize());
 app.use(passport.session());
 app.use(express.static(path.join(__dirname, 'public')));
 
+//routes
+var routes = require('./routes/index');
+var login = require('./routes/login');
+var auth = require('./routes/authReddit.js')(passport);
+
 app.use('/', routes);
 app.use('/login', login);
+app.use('/auth', auth);
 
-app.get('/auth/reddit', function(req, res, next){
-  req.session.state = crypto.randomBytes(32).toString('hex');
-  passport.authenticate('reddit', {
-    state: req.session.state,
-  })(req, res, next);
-});
-
-// GET /auth/reddit/callback
-//   Use passport.authenticate() as route middleware to authenticate the
-//   request.  If authentication fails, the user will be redirected back to the
-//   login page.  Otherwise, the primary route function function will be called,
-//   which, in this example, will redirect the user to the home page.
-app.get('/auth/reddit/callback', function(req, res, next){
-  // Check for origin via state token
-  if (req.query.state == req.session.state){
-    passport.authenticate('reddit', {
-      successRedirect: '/',
-      failureRedirect: '/login'
-    })(req, res, next);
-  }
-  else {
-    next( new Error(403) );
-  }
-});
-
-app.get('/logout', function(req, res){
+app.get('/logout', function(req, res) {
   req.logout();
   res.redirect('/');
 });
-
-function ensureAuthenticated(req, res, next) {
-  if (req.isAuthenticated()) { return next(); }
-  res.redirect('/login');
-}
-
-
-
-
-
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
