+++ ehackathon-server/middleware/ensureAuthenticated.js
@@ -1,0 +1,6 @@
+function ensureAuthenticated(req, res, next) {
+  if (req.isAuthenticated()) { return next(); }
+  res.redirect('/login');
+}
+
+module.exports = ensureAuthenticated;
