+++ ehackathon-server/views/login.jade
@@ -3,3 +3,4 @@ extends layout
 block content
   h1= title
   p Login page
+  a(href='/auth/reddit') Login w/ Reddit
